// src/generator/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
// ...
  export type Model = ReadonlyDeep<{
    name: string;
    dbName: string | null;
    fields: ModelField[];
    uniqueFields: string[][];
    uniqueIndexes: UniqueIndex[];
    // documentation?: string;
    primaryKey: PrimaryKey | null;
    // isGenerated?: boolean;
    // additional props:
    typeName: string;
    docs: string | undefined;
    plural: string | undefined;
    isOmitted: { output: boolean };
  }>;
// ...
  export type OutputType = ReadonlyDeep<{
    name: string;
    // fields: SchemaField[];
    // additional props:
    fields: OutputSchemaField[];
    typeName: string;
  }>;
// ...
  export type RelationModel = ReadonlyDeep<{
    model: Model;
    outputType: OutputType;
    relationFields: RelationField[];
    resolverName: string;
  }>;
  // additional type
  export type RelationField = ModelField &
    ReadonlyDeep<{
      outputTypeField: OutputSchemaField;
      argsTypeName: string | undefined;
    }>;
}